move source files out of src/ directory to the root directory
