+++ commands/domains.sh
