+++ commands/fsd.sh
