+++ internal/groups/groups.go
@@ -5,8 +5,8 @@ import (
 	"io/ioutil"
 	"log"
 
-	d "github.com/nattvara/dfb/src/internal/domains"
-	"github.com/nattvara/dfb/src/internal/paths"
+	d "github.com/nattvara/dfb/internal/domains"
+	"github.com/nattvara/dfb/internal/paths"
 )
 
 // FetchGroups reads and returns the groups stored on disk in dfp path
