+++ internal/stats/formatters.go
